fix(ng-version): logs, definitions, version, api
changes:
  - remove logging in gulpfile
  - update angular version in package.json
  - change several configuration names to be less verbose
  - update type definition files with new names

issues addressed:
  - related to issue #2

BREAKING CHANGE:
Upon updating to the latest version, the api has changed to make more sense and not require the repetetive "parallax" prefix to many of the properties of the parallaxConfig.  These names are likely to not change through the rest of the life of the project.

Rely upon the definitions to find the correct ones--the definition files haven't changed drastically, but also include some ts/jsdoc comments for better reading.  Further documentation will follow.